Fix for filters with more than 255 syscalls
It's not necessarily an issue right now, but rather a time bomb.
Jumps in BPF are limited to 255 insns. To work around this limit,
we insert a BPF_RET_TRACE insn every 255 syscalls we test
and adjust the jumps so that we always jump forward to the nearest
trace insn.

Signed-Off-By: Grzegorz Nosek <grzegorz.nosek@sysdig.com>