Simplify pdig attach
Instead of live patching the process when we attach to it,
let it run until the next syscall[*]. Then use the syscall
instruction that just trapped to call prctl() twice with
the parameters we need.

Also, store the SECCOMP filter on the target's stack
(below the red zone), instead of allocating and freeing memory.

* Note: this first syscall is not instrumented, but we don't have
  the process ID yet and we want to inject a system call
  instead of reading /proc, and can't do that until _after_
  the first system call

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>