Allow non-blocking attach
In order to attach to multiple processes, we need to wait
for each of them to execute a syscall. Since this can take
an arbitrary amount of time, we don't want to freeze all
other traced threads.

So, convert the whole thing to a state machine, where every*
waitpid() runs in the main loop.

Also, inject a getpid() system call instead of parsing /proc

* Note: waitpid()s done while single stepping are just executed
        inline, as they should finish quickly (we inject
        fast syscalls only)

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>