Allow attaching to whole process trees
Sadly, there's no system call we can use to find the relevant threads easily,
so we fall back to scanning /proc and building the whole parent-child
hierarchy. This is inherently racy, so we repeat the scan until we attach
to all the threads or our attempt counter runs out and we give up
(the workload is clone-heavy and apparently trying to evade us).

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>